+++ v2/internal/reconcilers/arm/arm_client_factory.go
@@ -1,0 +1,13 @@
+/*
+ * Copyright (c) Microsoft Corporation.
+ * Licensed under the MIT license.
+ */
+
+package arm
+
+import (
+	"github.com/Azure/azure-service-operator/v2/internal/genericarmclient"
+	"github.com/Azure/azure-service-operator/v2/pkg/genruntime"
+)
+
+type ARMClientFactory func(genruntime.MetaObject) *genericarmclient.GenericClient
