+++ v2/internal/testcommon/azure_be_provisioned_matcher.go
@@ -43,6 +43,9 @@ func (m *AzureBeProvisionedMatcher) Match(actual interface{}) (bool, error) {
 		return false, err
 	}
 
+	// The linter doesn't realize that we don't need to close the resp body because it's already done by the poller.
+	// Suppressing it as it is a false positive.
+	// nolint:bodyclose
 	_, err = poller.Poller.Poll(m.ctx)
 	if err != nil {
 		m.err = err
